add validation notebooks
