+++ models/Source/EMT/EMT_Ph3_SynchronGenerator3OrderVBR.cpp
@@ -66,10 +66,10 @@ void EMT::Ph3::SynchronGenerator3OrderVBR::calculateAuxiliarConstants() {
 
 void EMT::Ph3::SynchronGenerator3OrderVBR::stepInPerUnit() {
 
-	// calculate Eq_t at t=k
-	mEdq0_t(1,0) = mIdq0(0,0) * mLd_t + mVdq0(1,0);
-
 	if (mSimTime>0.0) {
+		// calculate Eq_t at t=k
+		mEdq0_t(1,0) = mIdq0(0,0) * mLd_t + mVdq0(1,0);
+
 		// calculate mechanical variables at t=k+1 with forward euler
 		mElecTorque = (mVdq0(0,0) * mIdq0(0,0) + mVdq0(1,0) * mIdq0(1,0));
 		mOmMech = mOmMech + mTimeStep * (1. / (2. * mH) * (mMechTorque - mElecTorque));
